Replaced original minecraft classes with disassembly.
Added disassembly of v4 server, moved v3 to folder

The client version changed from 3 to 4. I patched around
it but I can't tell what changed in the server code because
it's obfuscated completely randomly and all the class names
changed. For now things seem to work /ok/ but there's work
to be done for sure.