Updated the python matcher to be multithreaded.
Now completes in 25s, compared to 1m 20s for the bash version.
The initial pass is still an O(n^2) all pairs comparison between
the two sets but it's done in parallel by N-threads.

Then cases of duplicate matches are resolved by taking the best
match for the target (the bash script didn't do this).

Finally, the list is sorted by best match first.